src/agent/requirements.txt
openai>=1.60.0
azure-identity>=1.19.0
azure-storage-blob>=12.24.0
pyyaml>=6.0
